+++ src/codemagic/tools/_app_store_connect/arguments.py
@@ -360,7 +360,7 @@ class BuildArgument(cli.Argument):
         type=ResourceId,
         description='Alphanumeric ID value of the Beta Build Localization',
     )
-    LOCALE = cli.ArgumentProperties(
+    LOCALE_OPTIONAL = cli.ArgumentProperties(
         key='locale',
         flags=('--locale', '-l'),
         type=Locale,
@@ -373,6 +373,14 @@ class BuildArgument(cli.Argument):
             'required': False,
             'choices': list(Locale),
         },
+    )
+    LOCALE_DEFAULT = LOCALE_OPTIONAL.duplicate(
+        description=(
+            'The locale code name for displaying localized "What\'s new" content in TestFlight. '
+            "In case not provided, application's primary locale from test information is used instead. "
+            'Learn more from https://developer.apple.com/documentation/appstoreconnectapi/'
+            'betabuildlocalizationcreaterequest/data/attributes'
+        ),
     )
     WHATS_NEW = cli.ArgumentProperties(
         key='whats_new',